requirements.txt
aioredis==2.0.0
asgiref==3.4.1
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-filter==21.1
django-redis==5.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
factory-boy==3.2.1
Faker==9.8.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.4
MarkupSafe==2.0.1
packaging==21.3
psycopg2-binary==2.9.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
uWSGI==2.0.20
websockets==10.1
